cleanup

+++ Emulator/Components/Memory/C64MemoryBase.cpp
@@ -69,8 +69,8 @@ C64Memory::cacheInfo(MemInfo &result) const
         if (expansionPort.getGameLine()) result.bankMap |= 0x08;
         if (expansionPort.getExromLine()) result.bankMap |= 0x10;
 
-        for (int i = 0; i < 16; i++) result.peekSrc[i] = peekSrc[i];
-        for (int i = 0; i < 16; i++) result.vicPeekSrc[i] = vic.memSrc[i];
+        for (isize i = 0; i < 16; i++) result.peekSrc[i] = peekSrc[i];
+        for (isize i = 0; i < 16; i++) result.vicPeekSrc[i] = vic.memSrc[i];
     }
 }
